requirements.txt
fastapi==0.115.4
uvicorn[standard]
sqlalchemy==2.0.36
pandas==2.2.3
numpy
httpx==0.27.0
python-dotenv
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
